# openapi.yaml
openapi: 3.0.3
info:
  title: Expense Journal API
  version: 1.0.0
  description: API for tracking personal expenses with receipt photos

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                name:
                  type: string
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
        '409':
          description: User already exists

  /api/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  message:
                    type: string
        '401':
          description: Invalid credentials

  /api/auth/logout:
    post:
      summary: Logout user
      responses:
        '200':
          description: Logged out successfully

  /api/auth/session:
    get:
      summary: Get current session
      responses:
        '200':
          description: Session info
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  authenticated:
                    type: boolean

  /api/categories:
    get:
      summary: Get all expense categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

  /api/expenses:
    get:
      summary: Get user expenses
      parameters:
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '401':
          description: Unauthorized
    post:
      summary: Create new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, categoryId, date]
              properties:
                amount:
                  type: number
                description:
                  type: string
                categoryId:
                  type: string
                date:
                  type: string
                  format: date
                receiptUrl:
                  type: string
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '401':
          description: Unauthorized
        '400':
          description: Invalid input

  /api/expenses/{id}:
    put:
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                description:
                  type: string
                categoryId:
                  type: string
                date:
                  type: string
                  format: date
                receiptUrl:
                  type: string
      responses:
        '200':
          description: Expense updated
        '401':
          description: Unauthorized
        '404':
          description: Expense not found
    delete:
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Expense deleted
        '401':
          description: Unauthorized
        '404':
          description: Expense not found

  /api/expenses/summary:
    get:
      summary: Get expense summary for a month
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Monthly summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: number
                  byCategory:
                    type: array
                    items:
                      type: object
                      properties:
                        categoryId:
                          type: string
                        categoryName:
                          type: string
                        total:
                          type: number
                        count:
                          type: integer

  /api/upload:
    post:
      summary: Upload receipt image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
        '401':
          description: Unauthorized
        '400':
          description: Invalid file

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        color:
          type: string

    Expense:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date
        categoryId:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        receiptUrl:
          type: string
        createdAt:
          type: string
          format: date-time